requirements.txt
pandas==1.3.5
python-dotenv==0.21.1
requests==2.31.0
google-cloud-storage==2.19.0
google-cloud-bigquery==3.30.0
dbt-core==1.5.5
dbt-bigquery==1.5.2
streamlit==1.23.1
altair==5.0.1